ci: try gradle-nexus plugin
