+++ .github/workflows/release-manual.yaml
@@ -29,7 +29,7 @@ jobs:
 
       - name: Publish DDI Lifecycle lib on Maven Central
         run: |
-          ./gradlew :model:clean :model:publish
+          ./gradlew :model:clean :model:publishToSonatype :model:closeAndReleaseSonatypeStagingRepository
         env:
           OSSRH_USERNAME: ${{ secrets.OSSRH_USERNAME }}
           OSSRH_PASSWORD: ${{ secrets.OSSRH_TOKEN }}
